requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.8
pydantic==1.10.13
python-multipart==0.0.6
python-dotenv==1.0.0
